Rename: The mqtt_client_test
